refactor(settings): split the Dropbox tab into its own file
views_settings.rs sat at the repository's 1,000-line ceiling, leaving no
room for new settings. Move view_settings_dropbox_tab verbatim into
views_settings_dropbox.rs, mirroring the existing views_settings_perform
split, and share format_settings_bytes across the two.
